+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `pyrgg.engines.labeled_tree` module
 ## [2.0] - 2025-12-23
 ### Added
 - `pyrgg.engines.geometric_graph` module